Rewrite move engine, add game rule settings

+++ packages/wasm-checkers/src/util.rs
@@ -1,7 +1,7 @@
 pub struct BitUtil {}
 
 impl BitUtil {
-    pub fn is_set(number: u64, index: u64) -> bool {
+    pub fn is_set(number: u64, index: u16) -> bool {
         (number & (1_u64 << index)) != 0
     }
 }
